requirements.txt
flask
pandas
numpy
matplotlib
scikit-learn
scipy
tensorflow
keras
google-cloud-storage
flask-cors
prophet
statsmodels